feat(api): Cover main pages with mock data
Add fixed mock responses for the remaining main pages. Split the example data by resource type and verify sign-in through the local API.
