+++ apps/server/src/orpc/mock/app.test.ts
@@ -1,0 +1,47 @@
+import { createORPCClient } from "@orpc/client";
+import { RPCLink } from "@orpc/client/fetch";
+import type { RouterClient } from "@orpc/server";
+import { mockApp } from "./app";
+import { mockAccessToken, mockStats } from "./fixtures";
+import type { mockRouter } from "./router";
+
+function createMockClient(accessToken?: string) {
+  const link = new RPCLink({
+    url: "http://mock.local/rpc",
+    headers: accessToken
+      ? { authorization: `Bearer ${accessToken}` }
+      : undefined,
+    fetch: async (request) => mockApp.fetch(request),
+  });
+
+  // SAFETY: The client and test server use the same mock router.
+  return createORPCClient(link) as RouterClient<typeof mockRouter>;
+}
+
+describe("mock API over HTTP", () => {
+  it("handles requests before and after sign-in", async () => {
+    const anonymousClient = createMockClient();
+    await expect(anonymousClient.stats()).resolves.toEqual(mockStats);
+
+    const session = await anonymousClient.auth.login({
+      email: "qa@example.com",
+      password: "anything",
+    });
+    expect(session.accessToken).toBe(mockAccessToken);
+
+    const authenticatedClient = createMockClient(session.accessToken);
+    await expect(
+      authenticatedClient.notifications.list({ filter: "unread" }),
+    ).resolves.toMatchObject({
+      results: expect.arrayContaining([
+        expect.objectContaining({ read: false }),
+      ]),
+    });
+  });
+
+  it("exposes a health check", async () => {
+    const response = await mockApp.request("/_health");
+    expect(response.status).toBe(200);
+    await expect(response.json()).resolves.toEqual({ ok: true });
+  });
+});
