+++ apps/server/src/orpc/mock/fixtures/constants.ts
@@ -1,0 +1,1 @@
+export const timestamp = "2026-08-26T12:00:00.000Z";
