+++ apps/server/src/orpc/mock/fixtures/entities.ts
@@ -1,0 +1,173 @@
+import type { MockOutputs } from "../contract";
+import { timestamp } from "./constants";
+import { mockCountries, mockCountry, mockRegion, mockRegions } from "./places";
+
+type Entity = MockOutputs["entities"]["list"]["results"][number];
+
+// Producers and bottles
+export const mockEntity = {
+  id: 9201,
+  peatedId: "E9201",
+  name: "Lagavulin",
+  shortName: null,
+  type: ["brand", "distiller"],
+  kind: "distillery",
+  ownerId: null,
+  description: "An Islay distillery known for heavily peated single malt.",
+  descriptionSrc: "user",
+  yearEstablished: 1816,
+  website: "https://www.malts.com/en-row/distilleries/lagavulin",
+  country: mockCountry,
+  region: mockRegion,
+  address: null,
+  location: null,
+  totalTastings: 1200,
+  totalBottles: 84,
+  createdAt: timestamp,
+  updatedAt: timestamp,
+} satisfies Entity;
+
+export const mockEntities: Entity[] = [
+  mockEntity,
+  {
+    ...mockEntity,
+    id: 9202,
+    peatedId: "E9202",
+    name: "The Macallan",
+    shortName: "Macallan",
+    description:
+      "A Speyside distillery known for sherry-seasoned oak maturation.",
+    yearEstablished: 1824,
+    website: "https://www.themacallan.com",
+    region: mockRegions[1],
+    totalTastings: 2100,
+    totalBottles: 190,
+  },
+  {
+    ...mockEntity,
+    id: 9203,
+    peatedId: "E9203",
+    name: "Springbank",
+    shortName: null,
+    description:
+      "A Campbeltown distillery that malts, distills, matures, and bottles whisky on site.",
+    yearEstablished: 1828,
+    website: "https://www.springbank.scot",
+    region: mockRegions[3],
+    totalTastings: 980,
+    totalBottles: 120,
+  },
+  {
+    ...mockEntity,
+    id: 9204,
+    peatedId: "E9204",
+    name: "Buffalo Trace",
+    shortName: null,
+    description: "A Kentucky distillery that produces bourbon and rye whiskey.",
+    yearEstablished: 1775,
+    website: "https://www.buffalotracedistillery.com",
+    country: mockCountries[2],
+    region: mockRegions[4],
+    totalTastings: 1750,
+    totalBottles: 95,
+  },
+  {
+    ...mockEntity,
+    id: 9205,
+    peatedId: "E9205",
+    name: "Yamazaki",
+    shortName: null,
+    description: "Japan's first malt whisky distillery, founded near Kyoto.",
+    yearEstablished: 1923,
+    website: "https://house.suntory.com/yamazaki-whisky",
+    country: mockCountries[3],
+    region: null,
+    totalTastings: 1300,
+    totalBottles: 72,
+  },
+  {
+    ...mockEntity,
+    id: 9206,
+    peatedId: "E9206",
+    name: "Midleton",
+    shortName: null,
+    type: ["distiller"],
+    description:
+      "An Irish distillery that produces pot still and grain whiskey.",
+    yearEstablished: 1975,
+    website: "https://www.midletondistillerycollection.com",
+    country: mockCountries[1],
+    region: null,
+    totalTastings: 1500,
+    totalBottles: 160,
+  },
+  {
+    ...mockEntity,
+    id: 9207,
+    peatedId: "E9207",
+    name: "Redbreast",
+    shortName: null,
+    type: ["brand"],
+    kind: "brand",
+    description: "A range of Irish single pot still whiskey made at Midleton.",
+    yearEstablished: 1903,
+    website: "https://www.redbreastwhiskey.com",
+    country: mockCountries[1],
+    region: null,
+    totalTastings: 1100,
+    totalBottles: 34,
+  },
+  {
+    ...mockEntity,
+    id: 9208,
+    peatedId: "E9208",
+    name: "Gordon & MacPhail",
+    shortName: "G&M",
+    type: ["bottler"],
+    kind: "bottler",
+    description:
+      "An independent Scotch whisky bottler and maturation specialist.",
+    yearEstablished: 1895,
+    website: "https://www.gordonandmacphail.com",
+    region: mockRegions[1],
+    totalTastings: 760,
+    totalBottles: 1400,
+  },
+  {
+    ...mockEntity,
+    id: 9209,
+    peatedId: "E9209",
+    name: "Compass Box",
+    shortName: null,
+    type: ["brand", "bottler"],
+    kind: "blender",
+    description: "A Scotch whisky blending house founded in London.",
+    yearEstablished: 2000,
+    website: "https://www.compassboxwhisky.com",
+    region: null,
+    totalTastings: 890,
+    totalBottles: 75,
+  },
+  {
+    ...mockEntity,
+    id: 9210,
+    peatedId: "E9210",
+    name: "Diageo",
+    shortName: null,
+    type: [],
+    kind: "company",
+    description: "A global drinks company that owns several whisky brands.",
+    yearEstablished: 1997,
+    website: "https://www.diageo.com",
+    region: null,
+    totalTastings: 5200,
+    totalBottles: 620,
+  },
+];
+
+export const mockMacallanEntity = mockEntities[1]!;
+export const mockSpringbankEntity = mockEntities[2]!;
+export const mockBuffaloTraceEntity = mockEntities[3]!;
+export const mockYamazakiEntity = mockEntities[4]!;
+export const mockMidletonEntity = mockEntities[5]!;
+export const mockRedbreastEntity = mockEntities[6]!;
