+++ apps/server/src/orpc/mock/fixtures/index.ts
@@ -1,0 +1,21 @@
+export * from "./activity";
+export * from "./badges";
+export * from "./bottles";
+export * from "./catalog";
+export * from "./changes";
+export * from "./collections";
+export * from "./comments";
+export * from "./constants";
+export * from "./entities";
+export * from "./events";
+export * from "./flights";
+export * from "./friendships";
+export * from "./helpers";
+export * from "./notifications";
+export * from "./places";
+export * from "./prices";
+export * from "./reviews";
+export * from "./stats";
+export * from "./tastings";
+export * from "./user-stats";
+export * from "./users";
