+++ apps/server/src/orpc/mock/fixtures/reviews.ts
@@ -1,0 +1,189 @@
+import type { MockOutputs } from "../contract";
+import { mockBottle, mockBottles } from "./bottles";
+import { timestamp } from "./constants";
+
+type Review = MockOutputs["reviews"]["list"]["results"][number];
+
+export const mockReview = {
+  id: 9801,
+  name: mockBottle.fullName,
+  rating: 92,
+  url: "https://example.com/reviews/lagavulin-16",
+  site: {
+    id: 9802,
+    type: "whiskyadvocate",
+    name: "Whisky Advocate",
+    lastRunAt: timestamp,
+    nextRunAt: null,
+    runEvery: null,
+  },
+  article: {
+    title: "Lagavulin 16 Review",
+    publishedAt: timestamp,
+  },
+  reviewerName: "Mock Reviewer",
+  nativeScore: {
+    value: 92,
+    scale: 100,
+    display: "92",
+  },
+  summary: "Rich smoke, dried fruit, and maritime notes.",
+  bottle: mockBottle,
+  createdAt: timestamp,
+  updatedAt: timestamp,
+} satisfies Review;
+
+export const mockReviews = [
+  mockReview,
+  {
+    ...mockReview,
+    id: 9810,
+    rating: 88,
+    url: "https://example.com/reviews/lagavulin-8",
+    site: {
+      ...mockReview.site!,
+      id: 9811,
+      type: "whiskyfun",
+      name: "Whiskyfun",
+      runEvery: 1440,
+    },
+    article: {
+      title: "Lagavulin 8-year-old",
+      publishedAt: "2026-08-20T09:00:00.000Z",
+    },
+    reviewerName: "Serge Sample",
+    nativeScore: { value: 88, scale: 100, display: "88 points" },
+    summary: "Bright citrus and dry peat make this younger release lively.",
+    bottle: mockBottles[6],
+    createdAt: "2026-08-20T10:00:00.000Z",
+    updatedAt: "2026-08-20T10:00:00.000Z",
+  },
+  {
+    ...mockReview,
+    id: 9812,
+    name: mockBottles[1]!.fullName,
+    rating: 87,
+    url: "https://example.com/reviews/macallan-12-sherry-oak",
+    site: {
+      ...mockReview.site!,
+      id: 9813,
+      type: "whiskynotes",
+      name: "WhiskyNotes",
+      runEvery: 1440,
+    },
+    article: {
+      title: "Macallan 12 Sherry Oak Review",
+      publishedAt: "2026-08-18T14:00:00.000Z",
+    },
+    reviewerName: "Ruben Sample",
+    nativeScore: { value: 87, scale: 100, display: "87/100" },
+    summary: "Classic dried fruit and orange peel with polished oak.",
+    bottle: mockBottles[1],
+    createdAt: "2026-08-18T15:00:00.000Z",
+    updatedAt: "2026-08-18T15:00:00.000Z",
+  },
+  {
+    ...mockReview,
+    id: 9814,
+    name: mockBottles[3]!.fullName,
+    rating: 84,
+    url: "https://example.com/reviews/buffalo-trace-bourbon",
+    site: {
+      ...mockReview.site!,
+      id: 9815,
+      type: "bourbonculture",
+      name: "Bourbon Culture",
+      runEvery: 1440,
+    },
+    article: {
+      title: "Buffalo Trace Bourbon Review",
+      publishedAt: "2026-08-15T16:00:00.000Z",
+    },
+    reviewerName: "Bourbon Sample",
+    nativeScore: { value: 84, scale: 100, display: "84" },
+    summary: "An approachable bourbon with caramel, vanilla, and mild oak.",
+    bottle: mockBottles[3],
+    createdAt: "2026-08-15T17:00:00.000Z",
+    updatedAt: "2026-08-15T17:00:00.000Z",
+  },
+  {
+    ...mockReview,
+    id: 9816,
+    name: mockBottles[4]!.fullName,
+    rating: 91,
+    url: "https://example.com/reviews/yamazaki-12",
+    article: {
+      title: "Yamazaki 12-year-old Review",
+      publishedAt: "2026-08-12T11:00:00.000Z",
+    },
+    reviewerName: "Editorial Sample",
+    site: undefined,
+    nativeScore: { value: 91, scale: 100, display: "91" },
+    summary: "Layered orchard fruit, incense, and restrained oak.",
+    bottle: mockBottles[4],
+    createdAt: "2026-08-12T12:00:00.000Z",
+    updatedAt: "2026-08-12T12:00:00.000Z",
+  },
+  {
+    ...mockReview,
+    id: 9817,
+    name: "Mystery Islay Malt 18-year-old",
+    rating: 90,
+    url: "https://example.com/reviews/mystery-islay-18",
+    article: {
+      title: "Mystery Islay Malt Review",
+      publishedAt: "2026-08-10T12:00:00.000Z",
+    },
+    reviewerName: "Editorial Sample",
+    site: undefined,
+    nativeScore: { value: 90, scale: 100, display: "90" },
+    summary: "A mature smoky malt that has not yet been matched to a bottle.",
+    bottle: null,
+    createdAt: "2026-08-10T13:00:00.000Z",
+    updatedAt: "2026-08-10T13:00:00.000Z",
+  },
+  {
+    ...mockReview,
+    id: 9818,
+    rating: 89,
+    url: "https://example.com/reviews/lagavulin-16-dramface",
+    site: {
+      ...mockReview.site!,
+      id: 9819,
+      type: "dramface",
+      name: "Dramface",
+      runEvery: 1440,
+    },
+    article: {
+      title: "Lagavulin 16: The Islay Benchmark",
+      publishedAt: "2026-08-08T12:00:00.000Z",
+    },
+    reviewerName: "Dramface Sample",
+    nativeScore: { value: 8.9, scale: 10, display: "8.9/10" },
+    summary: "Balanced peat, fruit, and oak with an easy maritime character.",
+    createdAt: "2026-08-08T13:00:00.000Z",
+    updatedAt: "2026-08-08T13:00:00.000Z",
+  },
+  {
+    ...mockReview,
+    id: 9830,
+    rating: 90,
+    url: "https://example.com/reviews/lagavulin-16-words",
+    site: {
+      ...mockReview.site!,
+      id: 9831,
+      type: "wordsofwhisky",
+      name: "Words of Whisky",
+      runEvery: 1440,
+    },
+    article: {
+      title: "Revisiting Lagavulin 16-year-old",
+      publishedAt: "2026-08-05T12:00:00.000Z",
+    },
+    reviewerName: "Words Sample",
+    nativeScore: { value: 90, scale: 100, display: "90 points" },
+    summary: "Mature smoke, dried fruit, and sea salt remain well integrated.",
+    createdAt: "2026-08-05T13:00:00.000Z",
+    updatedAt: "2026-08-05T13:00:00.000Z",
+  },
+] satisfies Review[];
