+++ apps/server/src/orpc/mock/routes/notifications/list.ts
@@ -1,7 +1,7 @@
-import { mockNotifications } from "@peated/server/orpc/mock/fixtures";
+import { mockNotifications, mockPage } from "@peated/server/orpc/mock/fixtures";
 import { mockOS } from "@peated/server/orpc/mock/implementer";
 
-export default mockOS.notifications.count.handler(
+export default mockOS.notifications.list.handler(
   async ({ input, context, errors }) => {
     if (!context.user) {
       throw errors.UNAUTHORIZED();
@@ -11,6 +11,6 @@ export default mockOS.notifications.count.handler(
       input.filter === "unread"
         ? mockNotifications.filter((notification) => !notification.read)
         : mockNotifications;
-    return { count: notifications.length };
+    return mockPage(notifications, input.cursor, input.limit);
   },
 );
